Refactoring reorder function and added favicon
