+++ src/modules/addingFavico.js
@@ -1,0 +1,7 @@
+export default function setFavicons(favImg) {
+  const headTitle = document.querySelector('head');
+  const setFavicon = document.createElement('link');
+  setFavicon.setAttribute('rel', 'shortcut icon');
+  setFavicon.setAttribute('href', favImg);
+  headTitle.appendChild(setFavicon);
+}
