+++ src/style.css
@@ -81,7 +81,7 @@ body {
   color: var(--dark-color);
 }
 
-.dragging{
+.dragging {
   opacity: 0.5;
   cursor: grabbing;
 }
